[network] moving transport tests to their own file
