+++ network/src/lib.rs
@@ -21,6 +21,8 @@ mod peer;
 mod sink;
 pub mod transport;
 
+#[cfg(test)]
+mod transport_tests;
 #[cfg(not(any(feature = "testing", feature = "fuzzing")))]
 mod noise;
 #[cfg(any(feature = "testing", feature = "fuzzing"))]
